requirement.txt
asgiref==3.9.1
Django==5.2.6
django-axes==8.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
